templated workloads
