+++ src/job.ts
@@ -1,14 +1,13 @@
 import { ApiObject, Lazy, Duration } from 'cdk8s';
 import { Construct } from 'constructs';
-import { Workload, WorkloadProps } from './_workload';
+import { RestartPolicy, TemplatedWorkload, TemplatedWorkloadProps } from './_workload';
 import * as k8s from './imports/k8s';
-import { RestartPolicy } from './pod';
 
 
 /**
  * Properties for `Job`.
  */
-export interface JobProps extends WorkloadProps {
+export interface JobProps extends TemplatedWorkloadProps {
 
   /**
    * Specifies the duration the job may be active before the system tries to terminate it.
@@ -46,7 +45,7 @@ export interface JobProps extends WorkloadProps {
  * The Job object will start a new Pod if the first Pod fails or is deleted (for example due to a node hardware failure or a node reboot).
  * You can also use a Job to run multiple Pods in parallel.
  */
-export class Job extends Workload {
+export class Job extends TemplatedWorkload {
 
   /**
    * Duration before job is terminated. If undefined, there is no deadline.
